challenges/web/shopify/challenge/app/requirements.txt
selenium==4.18.1
flask==3.0.2
webdriver-manager==4.0.1
python_jwt==3.3.3
bcrypt==3.2.0
python-dotenv==0.19.2